+++ CHANGES.md
@@ -1,5 +1,7 @@
 # Change Log
 
+- Added SplitDirection property for display PointPrimitive and Billboard relative to the `Scene.splitPosition`. [#11982](https://github.com/CesiumGS/cesium/pull/11982)
+
 ### 1.119 - 2024-07-01
 
 #### @cesium/engine